Refactor: refactoring SearchBar including header style adjustment
