+++ src/Feedback/RestaurantForm.module.css
@@ -17,32 +17,6 @@
 .formLabel {
   margin-bottom: 5px;
 }
-/* 
-.input,
-.textarea,
-.checkboxContainer,
-.button {
-  padding: 8px;
-  margin-top: 2px;
-  border: 1px solid #ccc;
-  border-radius: 4px;
-}
-
-.checkboxContainer {
-  display: flex;
-  flex-direction: column;
-}
-
-.button {
-  background-color: #007bff;
-  color: white;
-  cursor: pointer;
-}
-
-.button:hover {
-  background-color: #0056b3;
-} */
-/* RestaurantForm.module.css 수정 */
 
 /* 기본적인 input 스타일 */
 .input,
