+++ src/Pureplate/Pureplate.module.css
@@ -212,18 +212,19 @@
   bottom: 47.95%;
   top: 50.68%;
 }
+
 .header {
   border-style: solid;
   border-color: transparent;
   border-width: 0.0625rem;
-  padding: 0.625rem 8.3125rem 0.625rem 8.3125rem;
+  padding: 0.625rem 0.625rem 0.625rem 0.625rem;
   display: flex;
   flex-direction: row;
   gap: 0rem;
   align-items: center;
   justify-content: flex-start;
-  width: 90rem;
-  /* height: 6.75rem; */
+  width: 100%;
+  height: 5rem;
   position: absolute;
   left: 0rem;
   top: 0rem;
@@ -232,29 +233,21 @@
 
 @media (max-width: 1000px) {
   .header {
-    padding: 0.625rem 8.3125rem 0.625rem 6.3125rem;
+    padding: 0.625rem 0.625rem 0.625rem 0.625rem;
   }
 }
 
 .purePlateIcon {
-  flex-shrink: 0;
-  width: 6rem;
-  height: 5rem;
-  position: absolute;
-  left: 0.875rem;
-  top: calc(50% - 2.5rem);
+  flex-shrink: 1;
+  width: 70%;
+  height: 70%;
+  min-width: 50%;
+  min-height: 50%;
+  position: relative;
+  align-items: center;
   object-fit: contain;
 }
 
-@media (max-width: 1000px) {
-  .purePlateIcon {
-    width: 4.8rem;
-    height: 4rem;
-    position: absolute;
-    left: 0.775rem;
-    top: calc(50% - 2rem);
-  }
-}
 .attributes {
   padding: 0.625rem;
   display: flex;
