--- src/Search/Search.module.css
@@ -1,53 +1,0 @@
-.search,
-.search * {
-  box-sizing: border-box;
-}
-.search {
-  width: 1.625rem;
-  height: 1.625rem;
-  position: relative;
-  overflow: visible;
-}
-
-.searchBar {
-  box-shadow: 0 1px 30px rgba(0, 0, 0, 0.06);
-  display: flex;
-  justify-content: flex-start;
-  overflow: hidden;
-  gap: 0px;
-  color: rgba(0, 0, 0, 0.5);
-  width: 400px;
-  white-space: nowrap;
-  text-align: center;
-}
-
-.searchInput {
-  font-family: Anek Bangla, sans-serif;
-  font-size: 16px;
-  border-radius: 8px 0 0 8px;
-  border: 1px solid rgba(211, 211, 211, 1);
-  background-color: rgba(255, 255, 255, 0.96);
-  padding: 12px 16px;
-  height: 44px; /* 입력 필드와 버튼의 높이를 동일하게 설정 */
-  flex: 1;
-}
-
-.searchButton {
-  border: none;
-  cursor: pointer;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  background-color: #4a7c59; /* 버튼 배경색 설정 */
-  height: 44px; /* 입력 필드와 버튼의 높이를 동일하게 설정 */
-  padding: 12px;
-  border-radius: 0 8px 8px 0; /* 버튼의 우측 모서리 둥글게 설정 */
-}
-
-.searchIcon {
-  width: 100%;
-  height: 100%;
-  align-items: center;
-  justify-content: center;
-  fill: white; /* 아이콘 색상 설정 */
-}
